Refactor and divide tests for LINE service functionality
