+++ tests/line/test_webhook.py
@@ -1,0 +1,70 @@
+"""Test LINE webhook endpoint functionality."""
+
+from collections.abc import Callable
+
+from fastapi.testclient import TestClient
+
+
+class TestLineWebhook:
+    """Test LINE webhook endpoint."""
+
+    def test_invalid_content_type(self, client: TestClient) -> None:
+        """Test webhook with invalid content type."""
+        response = client.post(
+            "/line/webhook",
+            content=b"{}",
+            headers={"X-Line-Signature": "test", "Content-Type": "text/plain"},
+        )
+        assert response.status_code == 400
+        assert response.json()["detail"] == "Invalid content type"
+
+    def test_webhook_processing_error(
+        self, client: TestClient, generate_line_signature: Callable[[bytes], str]
+    ) -> None:
+        """Test webhook processing error handling."""
+        body = b'{"invalid": "json"}'
+        signature = generate_line_signature(body)
+
+        response = client.post(
+            "/line/webhook",
+            content=body,
+            headers={"X-Line-Signature": signature, "Content-Type": "application/json"},
+        )
+        # Invalid webhook data should return 200 to stop LINE from retrying
+        assert response.status_code == 200
+        assert response.json() == {"message": "OK"}
+
+    def test_webhook_processing_success(
+        self, client: TestClient, generate_line_signature: Callable[[bytes], str]
+    ) -> None:
+        """Test successful webhook processing."""
+        # Valid LINE webhook payload with text message
+        body = (
+            b'{"events":[{"type":"message","replyToken":"test_token",'
+            b'"message":{"type":"text","text":"Hello"}}]}'
+        )
+        signature = generate_line_signature(body)
+
+        # Mock the webhook handler to not raise any exceptions
+        from unittest.mock import patch
+
+        with patch("app.line.service.webhook_handler.handle"):
+            response = client.post(
+                "/line/webhook",
+                content=body,
+                headers={"X-Line-Signature": signature, "Content-Type": "application/json"},
+            )
+            assert response.status_code == 200
+            assert response.json() == {"message": "OK"}
+
+    def test_webhook_invalid_signature(self, client: TestClient) -> None:
+        """Test webhook with invalid signature."""
+        body = b'{"events":[]}'
+
+        response = client.post(
+            "/line/webhook",
+            content=body,
+            headers={"X-Line-Signature": "invalid_signature", "Content-Type": "application/json"},
+        )
+        assert response.status_code == 400
+        assert response.json()["detail"] == "Invalid signature"
